refactor TrainForm into TrainPage
