+++ train-tracker/src/TrainPage.js
@@ -1,3 +1,5 @@
+import {useNavigate} from 'react-router-dom';
+
 function TrainPage(){
     const url = window.location.href;
     const split = url.split("/");
@@ -10,4 +12,24 @@ function TrainPage(){
         </div>
     )
 }
+
+export function TrainForm(){
+    const [selectedNumber, setSelectedNumber] = useState("");
+
+    const navigate = useNavigate();
+
+    function handleNumber(e){ setSelectedNumber(e.target.value); }
+
+    function search(event){
+        event.preventDefault();
+        navigate("/train/"+selectedNumber);
+    }
+
+    return (
+        <form onSubmit={search}>
+            <input className="select-box" value={selectedNumber} placeholder="Search by Number" onChange={handleNumber} type="number" min='1'></input>
+        </form>
+    )
+}
+
 export default TrainPage;
